+++ pysatNASA/instruments/methods/__init__.py
@@ -4,4 +4,5 @@ from pysatNASA.instruments.methods._cdf import CDF  # noqa F401
 from pysatNASA.instruments.methods import cdaweb  # noqa F401
 from pysatNASA.instruments.methods import cnofs  # noqa F401
 from pysatNASA.instruments.methods import de2  # noqa F401
+from pysatNASA.instruments.methods import general  # noqa F401
 from pysatNASA.instruments.methods import icon  # noqa F401