+++ pysatNASA/tests/test_instruments.py
@@ -6,73 +6,23 @@ Imports test methods from pysat.tests.instrument_test_class
 
 """
 
-import tempfile
-
-import pytest
-
-import pysat
 # Make sure to import your instrument library here
 import pysatNASA
 
 # Import the test classes from pysat
-from pysat.tests.instrument_test_class import InstTestClass
-from pysat.utils import generate_instrument_list
+from pysat.tests.classes.cls_instrument_library import InstLibTests
+
+# Tell the standard tests which instruments to run each test on.
+InstLibTests.initialize_test_package(InstLibTests,
+                                     inst_loc=pysatNASA.instruments)
 
 
-# Developers for instrument libraries should update the following line to
-# point to their own library package
-# e.g.,
-# instruments = generate_instrument_list(inst_loc=mypackage.instruments)
-instruments = generate_instrument_list(inst_loc=pysatNASA.instruments)
-
-# The following lines apply the custom instrument lists to each type of test
-method_list = [func for func in dir(InstTestClass)
-               if callable(getattr(InstTestClass, func))]
-# Search tests for iteration via pytestmark, update instrument list
-for method in method_list:
-    if hasattr(getattr(InstTestClass, method), 'pytestmark'):
-        # Get list of names of pytestmarks
-        Nargs = len(getattr(InstTestClass, method).pytestmark)
-        names = [getattr(InstTestClass, method).pytestmark[j].name
-                 for j in range(0, Nargs)]
-        # Add instruments from your library
-        if 'all_inst' in names:
-            mark = pytest.mark.parametrize("inst_name", instruments['names'])
-            getattr(InstTestClass, method).pytestmark.append(mark)
-        elif 'download' in names:
-            mark = pytest.mark.parametrize("inst_dict", instruments['download'])
-            getattr(InstTestClass, method).pytestmark.append(mark)
-        elif 'no_download' in names:
-            mark = pytest.mark.parametrize("inst_dict",
-                                           instruments['no_download'])
-            getattr(InstTestClass, method).pytestmark.append(mark)
-
-
-class TestInstruments(InstTestClass):
+class TestInstruments(InstLibTests):
     """Main class for instrument tests.
 
     Note
     ----
-    Uses class level setup and teardown so that all tests use the same
-    temporary directory. We do not want to geneate a new tempdir for each test,
-    as the load tests need to be the same as the download tests.
+    All standard tests, setup, and teardown inherited from the core pysat
+    instrument test class.
 
     """
-
-    def setup_class(self):
-        """Initialize the testing setup once before all tests are run."""
-        # Make sure to use a temporary directory so that the user's setup is not
-        # altered
-        self.tempdir = tempfile.TemporaryDirectory()
-        self.saved_path = pysat.params['data_dirs']
-        pysat.params.data['data_dirs'] = [self.tempdir.name]
-        # Developers for instrument libraries should update the following line
-        # to point to their own subpackage location, e.g.,
-        # self.inst_loc = mypackage.instruments
-        self.inst_loc = pysatNASA.instruments
-
-    def teardown_class(self):
-        """Clean up downloaded files and parameters from tests."""
-        pysat.params.data['data_dirs'] = self.saved_path
-        self.tempdir.cleanup()
-        del self.inst_loc, self.saved_path, self.tempdir