+++ CHANGELOG.md
@@ -2,7 +2,17 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/).
 
-## [0.0.3] - 2022-XX-XX
+## [0.0.4] - 2022-XX-XX
+* Update instrument tests with new test class
+* Bug Fixes
+  * Fixed a bug in metadata when loading GOLD Nmax data.
+  * Fixed a bug in user feedback for `methods.cdaweb.download`
+  * Fixed a bug in loading ICON IVM data (added multi_file_day = True)
+* Maintenance
+  * Reduce duplication of code in instrument modules
+  * Include flake8 linting of docstrings and style in Github Actions
+
+## [0.0.3] - 2022-05-18
 * Include flake8 linting of docstrings and style in Github Actions
 * Include Windows tests in Github Actions
 * Bug Fixes
@@ -16,7 +26,6 @@ This project adheres to [Semantic Versioning](https://semver.org/).
   * Update instrument template standards
   * Updated documentation style
   * Removed cap on cdflib
-
 
 ## [0.0.2] - 2021-06-07
 * Updated Instruments and routines to conform with changes made for pysat 3.0